Fix the renaming and using scss from inside recipes
