+++ package.json
@@ -26,8 +26,8 @@
     "@fortawesome/free-regular-svg-icons": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@rollup/plugin-json": "^6.0.1",
     "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-json": "^6.0.1",
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@types/react": "^18.2.38",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
@@ -49,6 +49,7 @@
     "framer-motion": "^10.16.5",
     "gulp": "^4.0.2",
     "gulp-livereload": "^4.0.2",
+    "gulp-replace": "^1.1.4",
     "gulp-sass": "^5.0.0",
     "gulp-sourcemaps": "^3.0.0",
     "gulp-strip-comments": "^2.6.0",
